RAG_Python_book/requirements.txt
streamlit
python-dotenv
langchain-community
langchain-groq
groq
